requirements (1).txt
streamlit==1.32.0
torch
torchvision
transformers
ultralytics
langchain
google-genai
langchain-google-genai
langchain-text-splitters
langchain-community
pypdf
Pillow
pandas
